+++ docs/3_Setting_Dev_Environments.md
@@ -9,7 +9,7 @@ docker exec -it lidar3d-RT bash
 - Install OpenPCDet based CenterPointPillar
 ``` shell
 cd ~/PerceptionRT
-python setup.py develop # sudo python setup.py develop
+sudo python setup.py develop # sudo python setup.py develop
 ```
 
 ## To Build Python module, you have to install and wrap the c++ to python API.
@@ -17,7 +17,8 @@ python setup.py develop # sudo python setup.py develop
 cd ~/
 git clone https://github.com/pybind/pybind11.git
 cd pybind11
-cmake .
+mkdir build && cd build
+cmake ..
 sudo make install
 
 pip install --upgrade pip
