+++ pcdet/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.6.0+762785e"
+__version__ = "0.6.0+51f2852"
